+++ lib/parse/src/parser.rs
@@ -1,421 +1,341 @@
-use super::operations::*;
-use crate::stream::Stream;
-use ast::{
-    types::{self, Type},
-    *,
-};
-use combine::{
-    attempt, choice, look_ahead, many, many1, none_of, one_of, optional,
-    parser::{
-        char::{alpha_num, char as character, digit, letter, newline, space, spaces, string},
-        combinator::{lazy, no_partial, not_followed_by},
-        regex::find,
-        sequence::between,
+use ast::{types::Type, *};
+use nom::{
+    branch::alt,
+    bytes::complete::tag,
+    character::complete::{
+        alpha1, alphanumeric0, alphanumeric1, char, digit1, hex_digit1, multispace0, multispace1,
+        none_of, one_of,
     },
-    sep_end_by, sep_end_by1, unexpected_any, value, Parser, Positioned,
+    combinator::{all_consuming, into, map, not, opt, peek, recognize, success, value, verify},
+    error::ParseError,
+    multi::{many0, many1, separated_list0, separated_list1},
+    number::complete::recognize_float,
+    sequence::{delimited, preceded, terminated, tuple},
+    IResult, InputLength, Parser,
 };
-use fnv::FnvHashSet;
-use once_cell::sync::Lazy;
+use nom_locate::LocatedSpan;
 use position::Position;
 
-static KEYWORDS: &[&str] = &[
+use crate::{
+    combinator::{separated_or_terminated_list0, separated_or_terminated_list1},
+    operations::{reduce_operations, SuffixOperator},
+};
+
+const KEYWORDS: &[&str] = &[
     "as", "else", "export", "for", "foreign", "if", "in", "import", "type",
 ];
 const OPERATOR_CHARACTERS: &str = "+-*/=<>&|!?";
+const OPERATOR_MODIFIERS: &str = "=";
 
-static BINARY_REGEX: Lazy<regex::Regex> = Lazy::new(|| regex::Regex::new(r"^0b[01]+").unwrap());
-static HEXADECIMAL_REGEX: Lazy<regex::Regex> =
-    Lazy::new(|| regex::Regex::new(r"^0x[0-9a-fA-F]+").unwrap());
-static DECIMAL_REGEX: Lazy<regex::Regex> =
-    Lazy::new(|| regex::Regex::new(r"^-?([1-9][0-9]*|0)(\.[0-9]+)?").unwrap());
-static STRING_CHARACTER_REGEX: Lazy<regex::Regex> =
-    Lazy::new(|| regex::Regex::new(r#"^[^\\"]"#).unwrap());
-static BYTE_CHARACTER_REGEX: Lazy<regex::Regex> =
-    Lazy::new(|| regex::Regex::new(r"^[0-9a-fA-F]{2}").unwrap());
+type Input<'a> = LocatedSpan<&'a str, &'a str>;
 
-pub fn module<'a>() -> impl Parser<Stream<'a>, Output = Module> {
-    (
-        position(),
-        blank(),
-        many(import()),
-        many(foreign_import()),
-        many(choice((
-            type_alias().map(TypeDefinition::from),
-            record_definition().map(TypeDefinition::from),
+pub fn input<'a>(source: &'a str, path: &'a str) -> Input<'a> {
+    LocatedSpan::new_extra(source, path)
+}
+
+pub fn module(input: Input) -> IResult<Input, Module> {
+    map(
+        all_consuming(tuple((
+            position,
+            many0(import),
+            many0(foreign_import),
+            many0(alt((into(type_alias), into(record_definition)))),
+            many0(definition),
+            blank,
         ))),
-        many(definition()),
-    )
-        .skip(eof())
-        .map(
-            |(position, _, imports, foreign_imports, type_definitions, definitions)| {
-                Module::new(
-                    imports,
-                    foreign_imports,
-                    type_definitions,
-                    definitions,
-                    position,
-                )
-            },
-        )
+        |(position, imports, foreign_imports, type_definitions, definitions, _)| {
+            Module::new(
+                imports,
+                foreign_imports,
+                type_definitions,
+                definitions,
+                position,
+            )
+        },
+    )(input)
 }
 
-pub fn comments<'a>() -> impl Parser<Stream<'a>, Output = Vec<Comment>> {
-    many(between(
-        spaces(),
-        spaces(),
-        choice((
-            comment().map(Some),
-            raw_string_literal().map(|_| None),
-            none_of("\"#".chars()).map(|_| None),
-        )),
-    ))
-    .skip(eof())
-    .map(|comments: Vec<_>| comments.into_iter().flatten().collect())
+pub fn comments(input: Input) -> IResult<Input, Vec<Comment>> {
+    map(
+        all_consuming(many0(tuple((
+            multispace0,
+            alt((
+                map(comment, Some),
+                map(raw_string_literal, |_| None),
+                map(none_of("\"#"), |_| None),
+            )),
+            multispace0,
+        )))),
+        |comments| {
+            comments
+                .into_iter()
+                .flat_map(|(_, comment, _)| comment)
+                .collect()
+        },
+    )(input)
 }
 
-fn import<'a>() -> impl Parser<Stream<'a>, Output = Import> {
-    (
-        attempt((
-            position(),
+fn import(input: Input) -> IResult<Input, Import> {
+    map(
+        tuple((
+            position,
             keyword("import"),
-            not_followed_by(keyword("foreign").with(value("foreign"))),
+            module_path,
+            opt(preceded(keyword("as"), identifier)),
+            opt(delimited(
+                sign("{"),
+                separated_or_terminated_list1(sign(","), unqualified_name),
+                sign("}"),
+            )),
         )),
-        module_path(),
-        optional(keyword("as").with(identifier())),
-        optional(between(
-            sign("{"),
-            sign("}"),
-            sep_end_by1(unqualified_name(), sign(",")),
-        )),
-    )
-        .map(|((position, _, _), path, prefix, names)| {
+        |(position, _, path, prefix, names)| {
             Import::new(path, prefix, names.unwrap_or_default(), position)
-        })
-        .expected("import statement")
+        },
+    )(input)
 }
 
-fn unqualified_name<'a>() -> impl Parser<Stream<'a>, Output = UnqualifiedName> {
-    token(attempt((position(), identifier())))
-        .map(|(position, identifier)| UnqualifiedName::new(identifier, position))
-        .expected("unqualified name")
+fn unqualified_name(input: Input) -> IResult<Input, UnqualifiedName> {
+    map(
+        token(tuple((position, identifier))),
+        |(position, identifier)| UnqualifiedName::new(identifier, position),
+    )(input)
 }
 
-fn module_path<'a>() -> impl Parser<Stream<'a>, Output = ModulePath> {
-    token(choice((
-        internal_module_path().map(ModulePath::from),
-        external_module_path().map(ModulePath::from),
-    )))
-    .expected("module path")
+fn module_path(input: Input) -> IResult<Input, ModulePath> {
+    token(alt((
+        into(internal_module_path),
+        into(external_module_path),
+    )))(input)
 }
 
-fn internal_module_path<'a>() -> impl Parser<Stream<'a>, Output = InternalModulePath> {
-    module_path_components(identifier()).map(InternalModulePath::new)
+fn internal_module_path(input: Input) -> IResult<Input, InternalModulePath> {
+    map(module_path_components(identifier), InternalModulePath::new)(input)
 }
 
-fn external_module_path<'a>() -> impl Parser<Stream<'a>, Output = ExternalModulePath> {
-    (
-        identifier(),
-        module_path_components(public_module_path_component()),
-    )
-        .map(|(package, components)| ExternalModulePath::new(package, components))
+fn external_module_path(input: Input) -> IResult<Input, ExternalModulePath> {
+    map(
+        tuple((
+            identifier,
+            module_path_components(public_module_path_component),
+        )),
+        |(package, components)| ExternalModulePath::new(package, components),
+    )(input)
 }
 
 fn module_path_components<'a>(
-    component: impl Parser<Stream<'a>, Output = String>,
-) -> impl Parser<Stream<'a>, Output = Vec<String>> {
-    many1(string(IDENTIFIER_SEPARATOR).with(component))
+    component: impl Parser<Input<'a>, String, nom::error::Error<Input<'a>>>,
+) -> impl FnMut(Input<'a>) -> IResult<Input, Vec<String>, nom::error::Error<Input<'a>>> {
+    many1(preceded(tag(IDENTIFIER_SEPARATOR), component))
 }
 
-fn public_module_path_component<'a>() -> impl Parser<Stream<'a>, Output = String> {
-    look_ahead(identifier().expected("public module path"))
-        .then(|name| {
-            if ast::analysis::is_name_public(&name) {
-                value(()).left()
-            } else {
-                unexpected_any("private module path").right()
-            }
-        })
-        .with(identifier())
+fn public_module_path_component(input: Input) -> IResult<Input, String> {
+    verify(identifier, |name| ast::analysis::is_name_public(&name))(input)
 }
 
-fn foreign_import<'a>() -> impl Parser<Stream<'a>, Output = ForeignImport> {
-    (
-        attempt(position().skip((keyword("import"), keyword("foreign")))),
-        optional(calling_convention()),
-        identifier(),
-        type_(),
-    )
-        .map(|(position, calling_convention, name, type_)| {
+fn foreign_import(input: Input) -> IResult<Input, ForeignImport> {
+    map(
+        tuple((
+            position,
+            keyword("import"),
+            keyword("foreign"),
+            opt(calling_convention),
+            identifier,
+            type_,
+        )),
+        |(position, _, _, calling_convention, name, type_)| {
             ForeignImport::new(
                 &name,
                 calling_convention.unwrap_or_default(),
                 type_,
                 position,
             )
-        })
-        .expected("foreign import statement")
-}
-
-fn calling_convention<'a>() -> impl Parser<Stream<'a>, Output = CallingConvention> {
-    string_literal()
-        .expected("calling convention")
-        .then(|string| {
-            if string.value() == "c" {
-                value(CallingConvention::C).left()
-            } else {
-                unexpected_any("unknown calling convention").right()
-            }
-        })
-}
-
-fn definition<'a>() -> impl Parser<Stream<'a>, Output = FunctionDefinition> {
-    (
-        optional(foreign_export()),
-        position(),
-        identifier(),
-        sign("="),
-        lambda(),
-    )
-        .map(|(foreign_export, position, name, _, lambda)| {
-            FunctionDefinition::new(name, lambda, foreign_export, position)
-        })
-        .expected("definition")
-}
-
-fn foreign_export<'a>() -> impl Parser<Stream<'a>, Output = ForeignExport> {
-    keyword("foreign")
-        .with(optional(calling_convention()))
-        .map(|calling_convention| ForeignExport::new(calling_convention.unwrap_or_default()))
-}
-
-fn record_definition<'a>() -> impl Parser<Stream<'a>, Output = RecordDefinition> {
-    (
-        attempt((position(), keyword("type"))),
-        identifier(),
-        sign("{"),
-        many((identifier(), type_())),
-        sign("}"),
-    )
-        .map(
-            |((position, _), name, _, fields, _): (_, _, _, Vec<_>, _)| {
-                RecordDefinition::new(
-                    name,
-                    fields
-                        .into_iter()
-                        .map(|(name, type_)| types::RecordField::new(name, type_))
-                        .collect(),
-                    position,
-                )
-            },
-        )
-        .expected("record definition")
-}
-
-fn type_alias<'a>() -> impl Parser<Stream<'a>, Output = TypeAlias> {
-    (
-        attempt((position(), keyword("type"), identifier(), sign("="))),
-        type_(),
-    )
-        .map(|((position, _, name, _), type_)| TypeAlias::new(name, type_, position))
-        .expected("type alias")
-}
-
-fn type_<'a>() -> impl Parser<Stream<'a>, Output = Type> {
-    lazy(|| no_partial(choice((function_type().map(Type::from), union_type()))))
-        .boxed()
-        .expected("type")
-}
-
-fn function_type<'a>() -> impl Parser<Stream<'a>, Output = types::Function> {
-    (
-        attempt(position().skip(sign("\\("))),
-        sep_end_by(type_(), sign(",")),
-        sign(")"),
-        type_(),
-    )
-        .map(|(position, arguments, _, result)| types::Function::new(arguments, result, position))
-        .expected("function type")
-}
-
-fn union_type<'a>() -> impl Parser<Stream<'a>, Output = Type> {
-    sep_end_by1(atomic_type(), sign("|"))
-        .map(|types: Vec<_>| {
-            types
-                .into_iter()
-                .reduce(|lhs, rhs| {
-                    types::Union::new(lhs.clone(), rhs, lhs.position().clone()).into()
-                })
-                .unwrap()
-        })
-        .expected("union type")
-}
-
-fn list_type<'a>() -> impl Parser<Stream<'a>, Output = types::List> {
-    (attempt(position().skip(sign("["))), type_(), sign("]"))
-        .map(|(position, element, _)| types::List::new(element, position))
-        .expected("list type")
-}
-
-fn map_type<'a>() -> impl Parser<Stream<'a>, Output = types::Map> {
-    (
-        attempt(position().skip(sign("{"))),
-        type_(),
-        sign(":"),
-        type_(),
-        sign("}"),
-    )
-        .map(|(position, key, _, value, _)| types::Map::new(key, value, position))
-        .expected("map type")
-}
-
-fn atomic_type<'a>() -> impl Parser<Stream<'a>, Output = Type> {
-    choice((
-        reference_type().map(Type::from),
-        list_type().map(Type::from),
-        map_type().map(Type::from),
-        between(sign("("), sign(")"), type_()),
-    ))
-}
-
-fn reference_type<'a>() -> impl Parser<Stream<'a>, Output = types::Reference> {
-    token(attempt((position(), qualified_identifier())))
-        .map(|(position, identifier)| types::Reference::new(identifier, position))
-        .expected("reference type")
-}
-
-fn block<'a>() -> impl Parser<Stream<'a>, Output = Block> {
-    (
-        position(),
-        between(sign("{"), sign("}"), many1(statement())),
-    )
-        .then(|(position, statements): (_, Vec<_>)| {
-            if let Some(statement) = statements.last() {
-                if statement.name().is_none() {
-                    value(Block::new(
-                        statements[..statements.len() - 1].to_vec(),
-                        statement.expression().clone(),
-                        position,
-                    ))
-                    .left()
-                } else {
-                    unexpected_any("end of block").right()
-                }
-            } else {
-                unexpected_any("end of block").right()
-            }
-        })
-        .expected("block")
-}
-
-fn statement<'a>() -> impl Parser<Stream<'a>, Output = Statement> {
-    choice((statement_with_result(), statement_without_result())).expected("statement")
-}
-
-fn statement_with_result<'a>() -> impl Parser<Stream<'a>, Output = Statement> {
-    (attempt((position(), identifier(), sign("="))), expression())
-        .map(|((position, name, _), expression)| Statement::new(Some(name), expression, position))
-}
-
-fn statement_without_result<'a>() -> impl Parser<Stream<'a>, Output = Statement> {
-    (position(), expression())
-        .map(|(position, expression)| Statement::new(None, expression, position))
-}
-
-fn expression<'a>() -> impl Parser<Stream<'a>, Output = Expression> {
-    lazy(|| no_partial(binary_operation_like()))
-        .boxed()
-        .expected("expression")
-}
-
-fn binary_operation_like<'a>() -> impl Parser<Stream<'a>, Output = Expression> {
-    (
-        prefix_operation_like(),
-        many(
-            (
-                attempt((position(), binary_operator())),
-                prefix_operation_like(),
-            )
-                .map(|((position, operator), expression)| (operator, expression, position)),
-        ),
-    )
-        .map(|(expression, pairs): (_, Vec<_>)| reduce_operations(expression, &pairs))
-}
-
-fn binary_operator<'a>() -> impl Parser<Stream<'a>, Output = BinaryOperator> {
-    choice((
-        concrete_binary_operator("+", BinaryOperator::Add),
-        concrete_binary_operator("-", BinaryOperator::Subtract),
-        concrete_binary_operator("*", BinaryOperator::Multiply),
-        concrete_binary_operator("/", BinaryOperator::Divide),
-        concrete_binary_operator("==", BinaryOperator::Equal),
-        concrete_binary_operator("!=", BinaryOperator::NotEqual),
-        concrete_binary_operator("<", BinaryOperator::LessThan),
-        concrete_binary_operator("<=", BinaryOperator::LessThanOrEqual),
-        concrete_binary_operator(">", BinaryOperator::GreaterThan),
-        concrete_binary_operator(">=", BinaryOperator::GreaterThanOrEqual),
-        concrete_binary_operator("&", BinaryOperator::And),
-        concrete_binary_operator("|", BinaryOperator::Or),
-    ))
-    .expected("binary operator")
-}
-
-fn concrete_binary_operator<'a>(
-    literal: &'static str,
-    operator: BinaryOperator,
-) -> impl Parser<Stream<'a>, Output = BinaryOperator> {
-    attempt(token(many1(one_of(OPERATOR_CHARACTERS.chars())).then(
-        move |parsed_literal: String| {
-            if parsed_literal == literal {
-                value(operator).left()
-            } else {
-                unexpected_any("unknown binary operator").right()
-            }
         },
-    )))
+    )(input)
 }
 
-fn prefix_operation_like<'a>() -> impl Parser<Stream<'a>, Output = Expression> {
-    lazy(|| {
-        no_partial(choice((
-            prefix_operation().map(Expression::from),
-            suffix_operation_like().map(Expression::from),
-        )))
-    })
-    .boxed()
+fn calling_convention(input: Input) -> IResult<Input, CallingConvention> {
+    value(
+        CallingConvention::C,
+        verify(string_literal, |string| string.value() == "c"),
+    )(input)
 }
 
-fn prefix_operation<'a>() -> impl Parser<Stream<'a>, Output = UnaryOperation> {
-    (
-        attempt((position(), prefix_operator())),
-        prefix_operation_like(),
-    )
-        .map(|((position, operator), expression)| {
-            UnaryOperation::new(operator, expression, position)
-        })
+fn definition(input: Input) -> IResult<Input, FunctionDefinition> {
+    map(
+        tuple((opt(foreign_export), position, identifier, sign("="), lambda)),
+        |(foreign_export, position, name, _, lambda)| {
+            FunctionDefinition::new(name, lambda, foreign_export, position)
+        },
+    )(input)
 }
 
-fn prefix_operator<'a>() -> impl Parser<Stream<'a>, Output = UnaryOperator> {
-    choice((concrete_prefix_operator("!", UnaryOperator::Not),)).expected("unary operator")
+fn foreign_export(input: Input) -> IResult<Input, ForeignExport> {
+    map(
+        preceded(keyword("foreign"), opt(calling_convention)),
+        |calling_convention| ForeignExport::new(calling_convention.unwrap_or_default()),
+    )(input)
 }
 
-fn concrete_prefix_operator<'a>(
-    literal: &'static str,
-    operator: UnaryOperator,
-) -> impl Parser<Stream<'a>, Output = UnaryOperator> {
-    token(
-        one_of(OPERATOR_CHARACTERS.chars()).then(move |parsed_literal: char| {
-            if parsed_literal.to_string() == literal {
-                value(operator).left()
-            } else {
-                unexpected_any("unknown unary operator").right()
-            }
-        }),
-    )
+fn record_definition(input: Input) -> IResult<Input, RecordDefinition> {
+    map(
+        tuple((
+            position,
+            keyword("type"),
+            identifier,
+            sign("{"),
+            many0(tuple((identifier, type_))),
+            sign("}"),
+        )),
+        |(position, _, name, _, fields, _)| {
+            RecordDefinition::new(
+                name,
+                fields
+                    .into_iter()
+                    .map(|(name, type_)| types::RecordField::new(name, type_))
+                    .collect(),
+                position,
+            )
+        },
+    )(input)
 }
 
-fn suffix_operation_like<'a>() -> impl Parser<Stream<'a>, Output = Expression> {
-    (atomic_expression(), many(suffix_operator())).map(
-        |(expression, suffix_operators): (_, Vec<_>)| {
+fn type_alias(input: Input) -> IResult<Input, TypeAlias> {
+    map(
+        tuple((position, keyword("type"), identifier, sign("="), type_)),
+        |(position, _, name, _, type_)| TypeAlias::new(name, type_, position),
+    )(input)
+}
+
+fn type_(input: Input) -> IResult<Input, Type> {
+    alt((into(function_type), union_type))(input)
+}
+
+fn function_type(input: Input) -> IResult<Input, types::Function> {
+    map(
+        tuple((
+            position,
+            sign("\\("),
+            separated_or_terminated_list0(sign(","), type_),
+            sign(")"),
+            type_,
+        )),
+        |(position, _, arguments, _, result)| types::Function::new(arguments, result, position),
+    )(input)
+}
+
+fn union_type(input: Input) -> IResult<Input, Type> {
+    map(separated_list1(sign("|"), atomic_type), |types| {
+        types
+            .into_iter()
+            .reduce(|lhs, rhs| types::Union::new(lhs.clone(), rhs, lhs.position().clone()).into())
+            .unwrap()
+    })(input)
+}
+
+fn list_type(input: Input) -> IResult<Input, types::List> {
+    map(
+        tuple((position, sign("["), type_, sign("]"))),
+        |(position, _, element, _)| types::List::new(element, position),
+    )(input)
+}
+
+fn map_type(input: Input) -> IResult<Input, types::Map> {
+    map(
+        tuple((position, sign("{"), type_, sign(":"), type_, sign("}"))),
+        |(position, _, key, _, value, _)| types::Map::new(key, value, position),
+    )(input)
+}
+
+fn atomic_type(input: Input) -> IResult<Input, Type> {
+    alt((
+        into(reference_type),
+        into(list_type),
+        into(map_type),
+        delimited(sign("("), type_, sign(")")),
+    ))(input)
+}
+
+fn reference_type(input: Input) -> IResult<Input, types::Reference> {
+    map(
+        tuple((position, token(qualified_identifier))),
+        |(position, identifier)| types::Reference::new(identifier, position),
+    )(input)
+}
+
+fn block(input: Input) -> IResult<Input, Block> {
+    map(
+        tuple((position, delimited(sign("{"), many1(statement), sign("}")))),
+        |(position, statements)| {
+            // TODO Validate the last expressions.
+            Block::new(
+                statements[..statements.len() - 1].to_vec(),
+                statements.last().unwrap().expression().clone(),
+                position,
+            )
+        },
+    )(input)
+}
+
+fn statement(input: Input) -> IResult<Input, Statement> {
+    map(
+        tuple((position, opt(terminated(identifier, sign("="))), expression)),
+        |(position, name, expression)| Statement::new(name, expression, position),
+    )(input)
+}
+
+fn expression(input: Input) -> IResult<Input, Expression> {
+    binary_operation_like(input)
+}
+
+fn binary_operation_like(input: Input) -> IResult<Input, Expression> {
+    map(
+        tuple((
+            prefix_operation_like,
+            many0(map(
+                tuple((position, binary_operator, prefix_operation_like)),
+                |(position, operator, expression)| (operator, expression, position),
+            )),
+        )),
+        |(expression, pairs)| reduce_operations(expression, &pairs),
+    )(input)
+}
+
+fn binary_operator(input: Input) -> IResult<Input, BinaryOperator> {
+    alt((
+        value(BinaryOperator::Add, sign("+")),
+        value(BinaryOperator::Subtract, sign("-")),
+        value(BinaryOperator::Multiply, sign("*")),
+        value(BinaryOperator::Divide, sign("/")),
+        value(BinaryOperator::Equal, sign("==")),
+        value(BinaryOperator::NotEqual, sign("!=")),
+        value(BinaryOperator::LessThanOrEqual, sign("<=")),
+        value(BinaryOperator::LessThan, sign("<")),
+        value(BinaryOperator::GreaterThanOrEqual, sign(">=")),
+        value(BinaryOperator::GreaterThan, sign(">")),
+        value(BinaryOperator::And, sign("&")),
+        value(BinaryOperator::Or, sign("|")),
+    ))(input)
+}
+
+fn prefix_operation_like(input: Input) -> IResult<Input, Expression> {
+    alt((into(prefix_operation), into(suffix_operation_like)))(input)
+}
+
+fn prefix_operation(input: Input) -> IResult<Input, UnaryOperation> {
+    map(
+        tuple((position, prefix_operator, prefix_operation_like)),
+        |(position, operator, expression)| UnaryOperation::new(operator, expression, position),
+    )(input)
+}
+
+fn prefix_operator(input: Input) -> IResult<Input, UnaryOperator> {
+    value(UnaryOperator::Not, sign("!"))(input)
+}
+
+fn suffix_operation_like(input: Input) -> IResult<Input, Expression> {
+    map(
+        tuple((atomic_expression, many0(suffix_operator))),
+        |(expression, suffix_operators)| {
             suffix_operators
                 .into_iter()
                 .fold(expression, |expression, operator| match operator {
@@ -430,462 +350,475 @@ fn suffix_operation_like<'a>() -> impl Parser<Stream<'a>, Output = Expression> {
                     }
                 })
         },
-    )
+    )(input)
 }
 
-fn suffix_operator<'a>() -> impl Parser<Stream<'a>, Output = SuffixOperator> {
-    choice((call_operator(), record_field_operator(), try_operator()))
+fn suffix_operator(input: Input) -> IResult<Input, SuffixOperator> {
+    alt((call_operator, record_field_operator, try_operator))(input)
 }
 
-fn call_operator<'a>() -> impl Parser<Stream<'a>, Output = SuffixOperator> {
-    (
-        attempt(position().skip(sign("("))),
-        sep_end_by(expression(), sign(",")).skip(sign(")")),
-    )
-        .map(|(position, arguments)| SuffixOperator::Call(arguments, position))
+fn call_operator(input: Input) -> IResult<Input, SuffixOperator> {
+    // Do not allow any space before parentheses.
+    map(
+        tuple((
+            peek(position),
+            tag("("),
+            separated_or_terminated_list0(sign(","), expression),
+            sign(")"),
+        )),
+        |(position, _, arguments, _)| SuffixOperator::Call(arguments, position),
+    )(input)
 }
 
-fn record_field_operator<'a>() -> impl Parser<Stream<'a>, Output = SuffixOperator> {
-    (attempt(position().skip(sign("."))), identifier())
-        .map(|(position, identifier)| SuffixOperator::RecordField(identifier, position))
+fn record_field_operator(input: Input) -> IResult<Input, SuffixOperator> {
+    map(
+        tuple((position, sign("."), identifier)),
+        |(position, _, identifier)| SuffixOperator::RecordField(identifier, position),
+    )(input)
 }
 
-fn try_operator<'a>() -> impl Parser<Stream<'a>, Output = SuffixOperator> {
-    position().skip(sign("?")).map(SuffixOperator::Try)
+fn try_operator(input: Input) -> IResult<Input, SuffixOperator> {
+    map(tuple((position, sign("?"))), |(position, _)| {
+        SuffixOperator::Try(position)
+    })(input)
 }
 
-fn atomic_expression<'a>() -> impl Parser<Stream<'a>, Output = Expression> {
-    lazy(|| {
-        no_partial(choice((
-            if_list().map(Expression::from),
-            if_map().map(Expression::from),
-            if_type().map(Expression::from),
-            if_().map(Expression::from),
-            lambda().map(Expression::from),
-            record().map(Expression::from),
-            list_comprehension().map(Expression::from),
-            list_literal().map(Expression::from),
-            map_literal().map(Expression::from),
-            number_literal().map(Expression::from),
-            string_literal().map(Expression::from),
-            variable().map(Expression::from),
-            between(sign("("), sign(")"), expression()),
-        )))
-    })
-    .boxed()
+fn atomic_expression(input: Input) -> IResult<Input, Expression> {
+    alt((
+        into(lambda),
+        into(if_list),
+        into(if_map),
+        into(if_type),
+        into(if_),
+        into(record),
+        into(list_comprehension),
+        into(list_literal),
+        into(map_literal),
+        into(number_literal),
+        into(string_literal),
+        into(variable),
+        delimited(sign("("), expression, sign(")")),
+    ))(input)
 }
 
-fn lambda<'a>() -> impl Parser<Stream<'a>, Output = Lambda> {
-    (
-        attempt(position().skip(sign("\\("))),
-        sep_end_by(argument(), sign(",")),
-        sign(")"),
-        type_(),
-        block(),
-    )
-        .map(|(position, arguments, _, result_type, body)| {
+fn lambda(input: Input) -> IResult<Input, Lambda> {
+    map(
+        tuple((
+            position,
+            sign("\\("),
+            separated_or_terminated_list0(sign(","), argument),
+            sign(")"),
+            type_,
+            block,
+        )),
+        |(position, _, arguments, _, result_type, body)| {
             Lambda::new(arguments, result_type, body, position)
-        })
-        .expected("function expression")
+        },
+    )(input)
 }
 
-fn argument<'a>() -> impl Parser<Stream<'a>, Output = Argument> {
-    (identifier(), type_()).map(|(name, type_)| Argument::new(name, type_))
+fn argument(input: Input) -> IResult<Input, Argument> {
+    map(tuple((identifier, type_)), |(name, type_)| {
+        Argument::new(name, type_)
+    })(input)
 }
 
-fn if_<'a>() -> impl Parser<Stream<'a>, Output = If> {
-    (
-        attempt(position().skip(keyword("if"))),
-        if_branch(),
-        many(attempt((keyword("else"), keyword("if"))).with(if_branch())),
-        keyword("else"),
-        block(),
-    )
-        .map(
-            |(position, first_branch, branches, _, else_block): (_, _, Vec<_>, _, _)| {
-                If::new(
-                    [first_branch].into_iter().chain(branches).collect(),
-                    else_block,
-                    position,
-                )
-            },
-        )
-        .expected("if expression")
-}
-
-fn if_branch<'a>() -> impl Parser<Stream<'a>, Output = IfBranch> {
-    (expression(), block()).map(|(expression, block)| IfBranch::new(expression, block))
-}
-
-fn if_list<'a>() -> impl Parser<Stream<'a>, Output = IfList> {
-    (
-        attempt(position().skip((keyword("if"), sign("[")))),
-        identifier(),
-        sign(","),
-        sign("..."),
-        identifier(),
-        sign("]"),
-        sign("="),
-        expression(),
-        block(),
-        keyword("else"),
-        block(),
-    )
-        .map(
-            |(position, first_name, _, _, rest_name, _, _, argument, then, _, else_)| {
-                IfList::new(argument, first_name, rest_name, then, else_, position)
-            },
-        )
-        .expected("if-list expression")
-}
-
-fn if_map<'a>() -> impl Parser<Stream<'a>, Output = IfMap> {
-    (
-        attempt((
-            position(),
+fn if_(input: Input) -> IResult<Input, If> {
+    map(
+        tuple((
+            position,
             keyword("if"),
-            identifier(),
-            sign("="),
-            expression(),
-            sign("["),
+            if_branch,
+            many0(preceded(tuple((keyword("else"), keyword("if"))), if_branch)),
+            keyword("else"),
+            block,
         )),
-        expression(),
-        sign("]"),
-        block(),
-        keyword("else"),
-        block(),
-    )
-        .map(|((position, _, name, _, map, _), key, _, then, _, else_)| {
-            IfMap::new(name, map, key, then, else_, position)
-        })
-        .expected("if-map expression")
+        |(position, _, first_branch, branches, _, else_block)| {
+            If::new(
+                [first_branch].into_iter().chain(branches).collect(),
+                else_block,
+                position,
+            )
+        },
+    )(input)
 }
 
-fn if_type<'a>() -> impl Parser<Stream<'a>, Output = IfType> {
-    (
-        attempt((position(), keyword("if"), identifier(), sign("="))),
-        expression(),
-        keyword("as"),
-        if_type_branch(),
-        many(attempt((keyword("else"), keyword("if"))).with(if_type_branch())),
-        optional(keyword("else").with(block())),
-    )
-        .map(
-            |((position, _, identifier, _), argument, _, first_branch, branches, else_): (
-                _,
-                _,
-                _,
-                _,
-                Vec<_>,
-                _,
-            )| {
-                IfType::new(
-                    identifier,
-                    argument,
-                    [first_branch].into_iter().chain(branches).collect(),
-                    else_,
+fn if_branch(input: Input) -> IResult<Input, IfBranch> {
+    map(tuple((expression, block)), |(expression, block)| {
+        IfBranch::new(expression, block)
+    })(input)
+}
+
+fn if_list(input: Input) -> IResult<Input, IfList> {
+    map(
+        tuple((
+            position,
+            keyword("if"),
+            sign("["),
+            identifier,
+            sign(","),
+            sign("..."),
+            identifier,
+            sign("]"),
+            sign("="),
+            expression,
+            block,
+            keyword("else"),
+            block,
+        )),
+        |(position, _, _, first_name, _, _, rest_name, _, _, argument, then, _, else_)| {
+            IfList::new(argument, first_name, rest_name, then, else_, position)
+        },
+    )(input)
+}
+
+fn if_map(input: Input) -> IResult<Input, IfMap> {
+    map(
+        tuple((
+            position,
+            keyword("if"),
+            identifier,
+            sign("="),
+            expression,
+            sign("["),
+            expression,
+            sign("]"),
+            block,
+            keyword("else"),
+            block,
+        )),
+        |(position, _, name, _, map, _, key, _, then, _, else_)| {
+            IfMap::new(name, map, key, then, else_, position)
+        },
+    )(input)
+}
+
+fn if_type(input: Input) -> IResult<Input, IfType> {
+    map(
+        tuple((
+            position,
+            keyword("if"),
+            identifier,
+            sign("="),
+            expression,
+            keyword("as"),
+            if_type_branch,
+            many0(preceded(
+                tuple((keyword("else"), keyword("if"))),
+                if_type_branch,
+            )),
+            opt(preceded(keyword("else"), block)),
+        )),
+        |(position, _, identifier, _, argument, _, first_branch, branches, else_)| {
+            IfType::new(
+                identifier,
+                argument,
+                [first_branch].into_iter().chain(branches).collect(),
+                else_,
+                position,
+            )
+        },
+    )(input)
+}
+
+fn if_type_branch(input: Input) -> IResult<Input, IfTypeBranch> {
+    map(tuple((type_, block)), |(type_, block)| {
+        IfTypeBranch::new(type_, block)
+    })(input)
+}
+
+fn record(input: Input) -> IResult<Input, Record> {
+    // TODO Validate duplicate fields.
+    map(
+        tuple((
+            position,
+            qualified_identifier,
+            sign("{"),
+            alt((
+                tuple((
+                    map(delimited(sign("..."), expression, sign(",")), Some),
+                    separated_or_terminated_list1(sign(","), record_field),
+                )),
+                tuple((
+                    success(None),
+                    separated_or_terminated_list0(sign(","), record_field),
+                )),
+            )),
+            sign("}"),
+        )),
+        |(position, name, _, (record, fields), _)| Record::new(name, record, fields, position),
+    )(input)
+}
+
+fn record_field(input: Input) -> IResult<Input, RecordField> {
+    map(
+        tuple((position, identifier, sign(":"), expression)),
+        |(position, name, _, expression)| RecordField::new(name, expression, position),
+    )(input)
+}
+
+fn number_literal(input: Input) -> IResult<Input, Number> {
+    map(
+        token(tuple((
+            position,
+            alt((binary_literal, hexadecimal_literal, decimal_literal)),
+            peek(not(digit1)),
+        ))),
+        |(position, number, _)| Number::new(number, position),
+    )(input)
+}
+
+fn binary_literal(input: Input) -> IResult<Input, NumberRepresentation> {
+    map(preceded(tag("0b"), many1(one_of("01"))), |characters| {
+        NumberRepresentation::Binary(String::from_iter(characters))
+    })(input)
+}
+
+fn hexadecimal_literal(input: Input) -> IResult<Input, NumberRepresentation> {
+    map(
+        preceded(tag("0x"), many1(one_of("0123456789abcdefABCDEF"))),
+        |characters| {
+            NumberRepresentation::Hexadecimal(String::from_iter(characters).to_lowercase())
+        },
+    )(input)
+}
+
+fn decimal_literal(input: Input) -> IResult<Input, NumberRepresentation> {
+    map(recognize_float, |characters: Input| {
+        NumberRepresentation::FloatingPoint(
+            String::from_utf8_lossy(characters.as_bytes()).to_string(),
+        )
+    })(input)
+}
+
+fn string_literal(input: Input) -> IResult<Input, ByteString> {
+    token(raw_string_literal)(input)
+}
+
+fn raw_string_literal(input: Input) -> IResult<Input, ByteString> {
+    map(
+        tuple((
+            position,
+            delimited(
+                char('"'),
+                many0(alt((
+                    recognize(none_of("\\\"")),
+                    tag("\\\\"),
+                    tag("\\\""),
+                    tag("\\n"),
+                    tag("\\r"),
+                    tag("\\t"),
+                    // TODO Limit a number of digits.
+                    recognize(tuple((tag("\\x"), hex_digit1))),
+                ))),
+                char('"'),
+            ),
+        )),
+        |(position, spans)| {
+            ByteString::new(
+                spans
+                    .iter()
+                    .map(|span| String::from_utf8_lossy(span.as_bytes()))
+                    .collect::<Vec<_>>()
+                    .concat(),
+                position,
+            )
+        },
+    )(input)
+}
+
+fn list_literal(input: Input) -> IResult<Input, List> {
+    map(
+        tuple((
+            position,
+            sign("["),
+            type_,
+            separated_or_terminated_list0(sign(","), list_element),
+            sign("]"),
+        )),
+        |(position, _, type_, elements, _)| List::new(type_, elements, position),
+    )(input)
+}
+
+fn list_element(input: Input) -> IResult<Input, ListElement> {
+    alt((
+        map(expression, ListElement::Single),
+        map(preceded(sign("..."), expression), ListElement::Multiple),
+    ))(input)
+}
+
+fn list_comprehension(input: Input) -> IResult<Input, Expression> {
+    map(
+        tuple((
+            position,
+            sign("["),
+            type_,
+            expression,
+            keyword("for"),
+            identifier,
+            opt(preceded(sign(","), identifier)),
+            keyword("in"),
+            expression,
+            sign("]"),
+        )),
+        |(position, _, type_, element, _, element_name, value_name, _, iterator, _)| {
+            if let Some(value_name) = value_name {
+                MapIterationComprehension::new(
+                    type_,
+                    element,
+                    element_name,
+                    value_name,
+                    iterator,
                     position,
                 )
-            },
-        )
-        .expected("if-type expression")
+                .into()
+            } else {
+                ListComprehension::new(type_, element, element_name, iterator, position).into()
+            }
+        },
+    )(input)
 }
 
-fn if_type_branch<'a>() -> impl Parser<Stream<'a>, Output = IfTypeBranch> {
-    (type_(), block()).map(|(type_, block)| IfTypeBranch::new(type_, block))
-}
-
-fn record<'a>() -> impl Parser<Stream<'a>, Output = Record> {
-    (
-        attempt((position(), qualified_identifier(), sign("{"))),
-        choice((
-            (
-                between(sign("..."), sign(","), expression()).map(Some),
-                sep_end_by1(record_field(), sign(",")),
-            ),
-            (value(None), sep_end_by(record_field(), sign(","))),
+fn map_literal(input: Input) -> IResult<Input, Map> {
+    map(
+        tuple((
+            position,
+            sign("{"),
+            type_,
+            sign(":"),
+            type_,
+            separated_or_terminated_list0(sign(","), map_element),
+            sign("}"),
         )),
-        sign("}"),
-    )
-        .then(|((position, name, _), (record, fields), _)| {
-            let fields: Vec<_> = fields;
-
-            if fields
-                .iter()
-                .map(|field| field.name())
-                .collect::<FnvHashSet<_>>()
-                .len()
-                == fields.len()
-            {
-                value(Record::new(name, record, fields, position)).left()
-            } else {
-                unexpected_any("duplicate keys in record literal").right()
-            }
-        })
-        .expected("record literal")
-}
-
-fn record_field<'a>() -> impl Parser<Stream<'a>, Output = RecordField> {
-    (attempt((position(), identifier())), sign(":"), expression())
-        .map(|((position, name), _, expression)| RecordField::new(name, expression, position))
-}
-
-fn number_literal<'a>() -> impl Parser<Stream<'a>, Output = Number> {
-    token(
-        attempt((
-            position(),
-            choice((binary_literal(), hexadecimal_literal(), decimal_literal())),
-        ))
-        .skip(not_followed_by(digit())),
-    )
-    .map(|(position, number)| Number::new(number, position))
-    .silent()
-    .expected("number literal")
-}
-
-fn binary_literal<'a>() -> impl Parser<Stream<'a>, Output = NumberRepresentation> {
-    let regex: &'static regex::Regex = &BINARY_REGEX;
-
-    find(regex).map(|string: &str| NumberRepresentation::Binary(string[2..].into()))
-}
-
-fn hexadecimal_literal<'a>() -> impl Parser<Stream<'a>, Output = NumberRepresentation> {
-    let regex: &'static regex::Regex = &HEXADECIMAL_REGEX;
-
-    find(regex).map(|string: &str| NumberRepresentation::Hexadecimal(string[2..].to_lowercase()))
-}
-
-fn decimal_literal<'a>() -> impl Parser<Stream<'a>, Output = NumberRepresentation> {
-    let regex: &'static regex::Regex = &DECIMAL_REGEX;
-
-    find(regex).map(|string: &str| NumberRepresentation::FloatingPoint(string.into()))
-}
-
-fn string_literal<'a>() -> impl Parser<Stream<'a>, Output = ByteString> {
-    token(raw_string_literal())
-}
-
-fn raw_string_literal<'a>() -> impl Parser<Stream<'a>, Output = ByteString> {
-    let string_regex: &'static regex::Regex = &STRING_CHARACTER_REGEX;
-    let byte_regex: &'static regex::Regex = &BYTE_CHARACTER_REGEX;
-
-    (
-        attempt(position().skip(character('"'))),
-        many(choice((
-            find(string_regex).map(String::from),
-            special_string_character("\\\\"),
-            special_string_character("\\\""),
-            special_string_character("\\n"),
-            special_string_character("\\r"),
-            special_string_character("\\t"),
-            (attempt(string("\\x")), find(byte_regex))
-                .map(|(prefix, byte)| prefix.to_owned() + byte),
-        ))),
-        character('"'),
-    )
-        .map(|(position, strings, _): (_, Vec<String>, _)| {
-            ByteString::new(strings.concat(), position)
-        })
-        .expected("string literal")
-}
-
-fn special_string_character<'a>(escape: &'static str) -> impl Parser<Stream<'a>, Output = String> {
-    attempt(string(escape)).map(String::from)
-}
-
-fn list_literal<'a>() -> impl Parser<Stream<'a>, Output = List> {
-    (
-        attempt(position().skip(sign("["))),
-        type_(),
-        sep_end_by(list_element(), sign(",")),
-        sign("]"),
-    )
-        .map(|(position, type_, elements, _)| List::new(type_, elements, position))
-        .expected("list literal")
-}
-
-fn list_element<'a>() -> impl Parser<Stream<'a>, Output = ListElement> {
-    choice((
-        expression().map(ListElement::Single),
-        sign("...").with(expression()).map(ListElement::Multiple),
-    ))
-}
-
-fn list_comprehension<'a>() -> impl Parser<Stream<'a>, Output = Expression> {
-    (
-        attempt((position(), sign("["), type_(), expression(), keyword("for"))),
-        identifier(),
-        optional(sign(",").with(identifier())),
-        keyword("in"),
-        expression(),
-        sign("]"),
-    )
-        .map(
-            |((position, _, type_, element, _), element_name, value_name, _, iterator, _)| {
-                if let Some(value_name) = value_name {
-                    MapIterationComprehension::new(
-                        type_,
-                        element,
-                        element_name,
-                        value_name,
-                        iterator,
-                        position,
-                    )
-                    .into()
-                } else {
-                    ListComprehension::new(type_, element, element_name, iterator, position).into()
-                }
-            },
-        )
-        .expected("list comprehension")
-}
-
-fn map_literal<'a>() -> impl Parser<Stream<'a>, Output = Map> {
-    (
-        attempt(position().skip(sign("{"))),
-        type_(),
-        sign(":"),
-        type_(),
-        sep_end_by(map_element(), sign(",")),
-        sign("}"),
-    )
-        .map(|(position, key_type, _, value_type, elements, _)| {
+        |(position, _, key_type, _, value_type, elements, _)| {
             Map::new(key_type, value_type, elements, position)
-        })
-        .expected("map literal")
+        },
+    )(input)
 }
 
-fn map_element<'a>() -> impl Parser<Stream<'a>, Output = MapElement> {
-    choice((
-        (
-            attempt((position(), expression()).skip(sign(":"))),
-            expression(),
-        )
-            .map(|((position, key), value)| MapEntry::new(key, value, position).into()),
-        sign("...").with(expression()).map(MapElement::Map),
-        expression().map(MapElement::Removal),
-    ))
+fn map_element(input: Input) -> IResult<Input, MapElement> {
+    alt((
+        map(
+            tuple((position, expression, sign(":"), expression)),
+            |(position, key, _, value)| MapEntry::new(key, value, position).into(),
+        ),
+        map(preceded(sign("..."), expression), MapElement::Map),
+        map(expression, MapElement::Removal),
+    ))(input)
 }
 
-fn variable<'a>() -> impl Parser<Stream<'a>, Output = Variable> {
-    token(attempt((position(), qualified_identifier())))
-        .map(|(position, identifier)| Variable::new(identifier, position))
-        .expected("variable")
+fn variable(input: Input) -> IResult<Input, Variable> {
+    map(
+        tuple((position, token(qualified_identifier))),
+        |(position, identifier)| Variable::new(identifier, position),
+    )(input)
 }
 
-fn qualified_identifier<'a>() -> impl Parser<Stream<'a>, Output = String> {
-    (
-        raw_identifier(),
-        optional(string(IDENTIFIER_SEPARATOR).with(raw_identifier())),
-    )
-        .map(|(former, latter)| {
-            latter
-                .map(|latter| [&former, IDENTIFIER_SEPARATOR, &latter].concat())
-                .unwrap_or(former)
-        })
-}
-
-fn identifier<'a>() -> impl Parser<Stream<'a>, Output = String> {
-    token(raw_identifier()).expected("identifier")
-}
-
-fn raw_identifier<'a>() -> impl Parser<Stream<'a>, Output = String> {
-    unchecked_identifier()
-        .then(|identifier| {
-            if KEYWORDS.contains(&identifier.as_str()) {
-                // TODO Fix those misuse of `unexpected_any` combinators.
-                // These lead to wrong positions in error messages.
-                // We use the `silent` method  below because its position is wrong anyway.
-                unexpected_any("keyword").left()
+fn qualified_identifier(input: Input) -> IResult<Input, String> {
+    map(
+        tuple((
+            raw_identifier,
+            opt(tuple((tag(IDENTIFIER_SEPARATOR), raw_identifier))),
+        )),
+        |(former, latter)| {
+            if let Some((_, latter)) = latter {
+                [&former, IDENTIFIER_SEPARATOR, &latter].concat()
             } else {
-                value(identifier).right()
+                former
             }
-        })
-        .silent()
+        },
+    )(input)
 }
 
-fn unchecked_identifier<'a>() -> impl Parser<Stream<'a>, Output = String> {
-    (
-        choice((letter(), combine::parser::char::char('_'))),
-        many(choice((alpha_num(), combine::parser::char::char('_')))),
-    )
-        .map(|(head, tail): (char, String)| [head.into(), tail].concat())
+fn identifier(input: Input) -> IResult<Input, String> {
+    token(raw_identifier)(input)
 }
 
-fn keyword<'a>(name: &'static str) -> impl Parser<Stream<'a>, Output = ()> {
+fn raw_identifier(input: Input) -> IResult<Input, String> {
+    verify(unchecked_identifier, |identifier: &str| {
+        !KEYWORDS.contains(&identifier)
+    })(input)
+}
+
+fn unchecked_identifier(input: Input) -> IResult<Input, String> {
+    map(
+        recognize(tuple((
+            alt((value((), alpha1::<Input, _>), value((), char('_')))),
+            many0(alt((value((), alphanumeric1), value((), char('_'))))),
+        ))),
+        |span| String::from_utf8_lossy(span.as_bytes()).to_string(),
+    )(input)
+}
+
+fn keyword(name: &'static str) -> impl FnMut(Input) -> IResult<Input, ()> {
     if !KEYWORDS.contains(&name) {
         unreachable!("undefined keyword");
     }
 
-    token(attempt(string(name)).skip(not_followed_by(choice((
-        alpha_num(),
-        combine::parser::char::char('_'),
-    )))))
-    .with(value(()))
-    .expected(name)
+    move |input| {
+        value(
+            (),
+            token(tuple((
+                tag(name),
+                peek(not(alt((value((), alphanumeric1), value((), char('_')))))),
+            ))),
+        )(input)
+    }
 }
 
-fn sign<'a>(sign: &'static str) -> impl Parser<Stream<'a>, Output = ()> {
-    let parser = string(sign);
+fn sign(sign: &'static str) -> impl Fn(Input) -> IResult<Input, ()> + Clone {
+    move |input| {
+        let parser = token(tag(sign));
 
-    token(
         if sign
             .chars()
             .any(|character| OPERATOR_CHARACTERS.contains(character))
         {
-            parser
-                .skip(not_followed_by(one_of(OPERATOR_CHARACTERS.chars())))
-                .left()
+            value((), tuple((parser, peek(not(one_of(OPERATOR_MODIFIERS))))))(input)
         } else {
-            parser.right()
-        },
-    )
-    .with(value(()))
-    .expected(sign)
+            value((), parser)(input)
+        }
+    }
 }
 
-fn token<'a, O, P: Parser<Stream<'a>, Output = O>>(p: P) -> impl Parser<Stream<'a>, Output = O> {
-    p.skip(blank())
+fn token<'a, O>(
+    mut parser: impl Parser<Input<'a>, O, nom::error::Error<Input<'a>>>,
+) -> impl FnMut(Input<'a>) -> IResult<Input, O, nom::error::Error<Input<'a>>> {
+    move |input| {
+        let (input, _) = blank(input)?;
+
+        parser.parse(input)
+    }
 }
 
-fn position<'a>() -> impl Parser<Stream<'a>, Output = Position> {
-    value(()).map_input(|_, stream: &mut Stream<'a>| {
-        let position = stream.position();
+fn blank(input: Input) -> IResult<Input, ()> {
+    value((), many0(alt((value((), multispace1), value((), comment)))))(input)
+}
 
+fn comment(input: Input) -> IResult<Input, Comment> {
+    map(
+        tuple((position, tag("#"), many0(none_of("\n\r")))),
+        |(position, _, characters)| Comment::new(String::from_iter(characters), position),
+    )(input)
+}
+
+fn position(input: Input) -> IResult<Input, Position> {
+    let (input, _) = multispace0(input)?;
+
+    Ok((
+        input,
         Position::new(
-            &stream.0.state.path,
-            position.line as usize,
-            position.column as usize,
-            stream.0.state.lines[position.line as usize - 1],
-        )
-    })
-}
-
-fn eof<'a>() -> impl Parser<Stream<'a>, Output = ()> {
-    combine::eof().expected("end of file")
-}
-
-fn blank<'a>() -> impl Parser<Stream<'a>, Output = ()> {
-    many::<Vec<_>, _, _>(choice((space().with(value(())), comment().with(value(())))))
-        .with(value(()))
-}
-
-fn comment<'a>() -> impl Parser<Stream<'a>, Output = Comment> {
-    (
-        attempt((position(), string("#"))),
-        many::<Vec<_>, _, _>(none_of("\n".chars())),
-    )
-        .map(|((position, _), string)| {
-            Comment::new(string.into_iter().collect::<String>().trim_end(), position)
-        })
-        .skip(choice((newline().with(value(())), eof())))
-        .expected("comment")
+            input.extra,
+            input.location_line() as usize,
+            input.get_column(),
+            String::from_utf8_lossy(input.get_line_beginning()),
+        ),
+    ))
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{error::ParseError, stream::stream};
     use indoc::indoc;
     use position::test::PositionFake;
     use pretty_assertions::assert_eq;
@@ -897,19 +830,19 @@ mod tests {
         #[test]
         fn parse_module() {
             assert_eq!(
-                module().parse(stream("", "")).unwrap().0,
+                module(input("", "")).unwrap().1,
                 Module::new(vec![], vec![], vec![], vec![], Position::fake())
             );
             assert_eq!(
-                module().parse(stream(" ", "")).unwrap().0,
+                module(input(" ", "")).unwrap().1,
                 Module::new(vec![], vec![], vec![], vec![], Position::fake())
             );
             assert_eq!(
-                module().parse(stream("\n", "")).unwrap().0,
+                module(input("\n", "")).unwrap().1,
                 Module::new(vec![], vec![], vec![], vec![], Position::fake())
             );
             assert_eq!(
-                module().parse(stream("import Foo'Bar", "")).unwrap().0,
+                module(input("import Foo'Bar", "")).unwrap().1,
                 Module::new(
                     vec![Import::new(
                         ExternalModulePath::new("Foo", vec!["Bar".into()]),
@@ -924,7 +857,7 @@ mod tests {
                 )
             );
             assert_eq!(
-                module().parse(stream("type foo = number", "")).unwrap().0,
+                module(input("type foo = number", "")).unwrap().1,
                 Module::new(
                     vec![],
                     vec![],
@@ -939,10 +872,7 @@ mod tests {
                 )
             );
             assert_eq!(
-                module()
-                    .parse(stream("x=\\(x number)number{42}", ""))
-                    .unwrap()
-                    .0,
+                module(input("x=\\(x number)number{42}", "")).unwrap().1,
                 Module::new(
                     vec![],
                     vec![],
@@ -972,13 +902,12 @@ mod tests {
                 )
             );
             assert_eq!(
-                module()
-                    .parse(stream(
-                        "x=\\(x number)number{42}y=\\(y number)number{42}",
-                        ""
-                    ))
-                    .unwrap()
-                    .0,
+                module(input(
+                    "x=\\(x number)number{42}y=\\(y number)number{42}",
+                    ""
+                ))
+                .unwrap()
+                .1,
                 Module::new(
                     vec![],
                     vec![],
@@ -1035,10 +964,9 @@ mod tests {
         #[test]
         fn parse_import_foreign_after_import() {
             assert_eq!(
-                module()
-                    .parse(stream("import Foo'Bar import foreign foo \\() number", ""))
+                module(input("import Foo'Bar import foreign foo \\() number", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 Module::new(
                     vec![Import::new(
                         ExternalModulePath::new("Foo", vec!["Bar".into()]),
@@ -1066,10 +994,9 @@ mod tests {
         #[test]
         fn parse_record_definition_after_type_alias() {
             assert_eq!(
-                module()
-                    .parse(stream("type foo = number type bar {}", ""))
+                module(input("type foo = number type bar {}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 Module::new(
                     vec![],
                     vec![],
@@ -1096,7 +1023,7 @@ mod tests {
         #[test]
         fn parse_import() {
             assert_eq!(
-                import().parse(stream("import 'Foo", "")).unwrap().0,
+                import(input("import 'Foo", "")).unwrap().1,
                 Import::new(
                     InternalModulePath::new(vec!["Foo".into()]),
                     None,
@@ -1105,7 +1032,7 @@ mod tests {
                 ),
             );
             assert_eq!(
-                import().parse(stream("import Foo'Bar", "")).unwrap().0,
+                import(input("import Foo'Bar", "")).unwrap().1,
                 Import::new(
                     ExternalModulePath::new("Foo", vec!["Bar".into()]),
                     None,
@@ -1118,7 +1045,7 @@ mod tests {
         #[test]
         fn parse_import_with_custom_prefix() {
             assert_eq!(
-                import().parse(stream("import 'Foo as foo", "")).unwrap().0,
+                import(input("import 'Foo as foo", "")).unwrap().1,
                 Import::new(
                     InternalModulePath::new(vec!["Foo".into()]),
                     Some("foo".into()),
@@ -1131,7 +1058,7 @@ mod tests {
         #[test]
         fn parse_unqualified_import() {
             assert_eq!(
-                import().parse(stream("import 'Foo { Foo }", "")).unwrap().0,
+                import(input("import 'Foo { Foo }", "")).unwrap().1,
                 Import::new(
                     InternalModulePath::new(vec!["Foo".into()]),
                     None,
@@ -1144,10 +1071,7 @@ mod tests {
         #[test]
         fn parse_unqualified_import_with_multiple_identifiers() {
             assert_eq!(
-                import()
-                    .parse(stream("import 'Foo { Foo, Bar }", ""))
-                    .unwrap()
-                    .0,
+                import(input("import 'Foo { Foo, Bar }", "")).unwrap().1,
                 Import::new(
                     InternalModulePath::new(vec!["Foo".into()]),
                     None,
@@ -1162,73 +1086,65 @@ mod tests {
 
         #[test]
         fn parse_module_path() {
-            assert!(module_path().parse(stream("", "")).is_err());
+            assert!(module_path(input("", "")).is_err());
             assert_eq!(
-                module_path().parse(stream("'Foo", "")).unwrap().0,
+                module_path(input("'Foo", "")).unwrap().1,
                 InternalModulePath::new(vec!["Foo".into()]).into(),
             );
             assert_eq!(
-                module_path().parse(stream("Foo'Bar", "")).unwrap().0,
+                module_path(input("Foo'Bar", "")).unwrap().1,
                 ExternalModulePath::new("Foo", vec!["Bar".into()]).into(),
             );
         }
 
         #[test]
         fn parse_internal_module_path() {
-            assert!(internal_module_path().parse(stream("", "")).is_err());
+            assert!(internal_module_path(input("", "")).is_err());
             assert_eq!(
-                internal_module_path().parse(stream("'Foo", "")).unwrap().0,
+                internal_module_path(input("'Foo", "")).unwrap().1,
                 InternalModulePath::new(vec!["Foo".into()]),
             );
             assert_eq!(
-                internal_module_path()
-                    .parse(stream("'Foo'Bar", ""))
-                    .unwrap()
-                    .0,
+                internal_module_path(input("'Foo'Bar", "")).unwrap().1,
                 InternalModulePath::new(vec!["Foo".into(), "Bar".into()]),
             );
         }
 
         #[test]
         fn parse_external_module_path() {
-            assert!(external_module_path().parse(stream("", "")).is_err());
+            assert!(external_module_path(input("", "")).is_err());
             assert_eq!(
-                external_module_path()
-                    .parse(stream("Foo'Bar", ""))
-                    .unwrap()
-                    .0,
+                external_module_path(input("Foo'Bar", "")).unwrap().1,
                 ExternalModulePath::new("Foo", vec!["Bar".into()]),
             );
         }
 
-        #[test]
-        fn fail_to_parse_private_external_module_file() {
-            let source = "Foo'bar";
+        // TODO
+        // #[test]
+        // fn fail_to_parse_private_external_module_file() {
+        //     let source = "Foo'bar";
 
-            insta::assert_debug_snapshot!(external_module_path()
-                .parse(stream(source, ""))
-                .map_err(|error| ParseError::new(source, "", error))
-                .err());
-        }
+        //     insta::assert_debug_snapshot!(external_module_path(input(source, ""))
+        //         .map_err(|error| ParseError::new(source, "", error))
+        //         .err());
+        // }
 
-        #[test]
-        fn fail_to_parse_private_external_module_directory() {
-            let source = "Foo'bar'Baz";
+        // #[test]
+        // fn fail_to_parse_private_external_module_directory() {
+        //     let source = "Foo'bar'Baz";
 
-            insta::assert_debug_snapshot!(external_module_path()
-                .parse(stream(source, ""))
-                .map_err(|error| ParseError::new(source, "", error))
-                .err());
-        }
+        //     insta::assert_debug_snapshot!(external_module_path(input(source, ""))
+        //         .map_err(|error| ParseError::new(source, "", error))
+        //         .err());
+        // }
     }
 
     #[test]
     fn parse_foreign_import() {
         assert_eq!(
-            foreign_import()
-                .parse(stream("import foreign foo \\(number) number", ""))
+            foreign_import(input("import foreign foo \\(number) number", ""))
                 .unwrap()
-                .0,
+                .1,
             ForeignImport::new(
                 "foo",
                 CallingConvention::Native,
@@ -1242,10 +1158,9 @@ mod tests {
         );
 
         assert_eq!(
-            foreign_import()
-                .parse(stream("import foreign \"c\" foo \\(number) number", ""))
+            foreign_import(input("import foreign \"c\" foo \\(number) number", ""))
                 .unwrap()
-                .0,
+                .1,
             ForeignImport::new(
                 "foo",
                 CallingConvention::C,
@@ -1266,10 +1181,7 @@ mod tests {
         #[test]
         fn parse() {
             assert_eq!(
-                definition()
-                    .parse(stream("x=\\(x number)number{42}", ""))
-                    .unwrap()
-                    .0,
+                definition(input("x=\\(x number)number{42}", "")).unwrap().1,
                 FunctionDefinition::new(
                     "x",
                     Lambda::new(
@@ -1297,10 +1209,9 @@ mod tests {
         #[test]
         fn parse_foreign_definition() {
             assert_eq!(
-                definition()
-                    .parse(stream("foreign x=\\(x number)number{42}", ""))
+                definition(input("foreign x=\\(x number)number{42}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 FunctionDefinition::new(
                     "x",
                     Lambda::new(
@@ -1328,10 +1239,9 @@ mod tests {
         #[test]
         fn parse_foreign_definition_with_c_calling_convention() {
             assert_eq!(
-                definition()
-                    .parse(stream("foreign \"c\" x=\\(x number)number{42}", ""))
+                definition(input("foreign \"c\" x=\\(x number)number{42}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 FunctionDefinition::new(
                     "x",
                     Lambda::new(
@@ -1359,10 +1269,9 @@ mod tests {
         #[test]
         fn parse_keyword_like_name() {
             assert_eq!(
-                definition()
-                    .parse(stream("importA = \\() number { 42 }", ""))
+                definition(input("importA = \\() number { 42 }", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 FunctionDefinition::new(
                     "importA",
                     Lambda::new(
@@ -1421,10 +1330,7 @@ mod tests {
                 ),
             ),
         ] {
-            assert_eq!(
-                &record_definition().parse(stream(source, "")).unwrap().0,
-                expected
-            );
+            assert_eq!(&record_definition(input(source, "")).unwrap().1, expected);
         }
     }
 
@@ -1460,39 +1366,39 @@ mod tests {
                 ),
             ),
         ] {
-            assert_eq!(&type_alias().parse(stream(source, "")).unwrap().0, expected);
+            assert_eq!(&type_alias(input(source, "")).unwrap().1, expected);
         }
     }
 
-    mod types_ {
+    mod type_ {
         use super::*;
         use pretty_assertions::assert_eq;
 
         #[test]
         fn parse_type() {
-            assert!(type_().parse(stream("", "")).is_err());
+            assert!(type_(input("", "")).is_err());
             assert_eq!(
-                type_().parse(stream("boolean", "")).unwrap().0,
+                type_(input("boolean", "")).unwrap().1,
                 types::Reference::new("boolean", Position::fake()).into()
             );
             assert_eq!(
-                type_().parse(stream("none", "")).unwrap().0,
+                type_(input("none", "")).unwrap().1,
                 types::Reference::new("none", Position::fake()).into()
             );
             assert_eq!(
-                type_().parse(stream("number", "")).unwrap().0,
+                type_(input("number", "")).unwrap().1,
                 types::Reference::new("number", Position::fake()).into()
             );
             assert_eq!(
-                type_().parse(stream("Foo", "")).unwrap().0,
+                type_(input("Foo", "")).unwrap().1,
                 types::Reference::new("Foo", Position::fake()).into()
             );
             assert_eq!(
-                type_().parse(stream("Foo'Bar", "")).unwrap().0,
+                type_(input("Foo'Bar", "")).unwrap().1,
                 types::Reference::new("Foo'Bar", Position::fake()).into()
             );
             assert_eq!(
-                type_().parse(stream("\\(number)number", "")).unwrap().0,
+                type_(input("\\(number)number", "")).unwrap().1,
                 types::Function::new(
                     vec![types::Reference::new("number", Position::fake()).into()],
                     types::Reference::new("number", Position::fake()),
@@ -1501,10 +1407,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                type_()
-                    .parse(stream("\\(number,number)number", ""))
-                    .unwrap()
-                    .0,
+                type_(input("\\(number,number)number", "")).unwrap().1,
                 types::Function::new(
                     vec![
                         types::Reference::new("number", Position::fake()).into(),
@@ -1516,10 +1419,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                type_()
-                    .parse(stream("\\(\\(number)number)number", ""))
-                    .unwrap()
-                    .0,
+                type_(input("\\(\\(number)number)number", "")).unwrap().1,
                 types::Function::new(
                     vec![types::Function::new(
                         vec![types::Reference::new("number", Position::fake()).into()],
@@ -1533,7 +1433,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                type_().parse(stream("number|none", "")).unwrap().0,
+                type_(input("number|none", "")).unwrap().1,
                 types::Union::new(
                     types::Reference::new("number", Position::fake()),
                     types::Reference::new("none", Position::fake()),
@@ -1542,7 +1442,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                type_().parse(stream("boolean|number|none", "")).unwrap().0,
+                type_(input("boolean|number|none", "")).unwrap().1,
                 types::Union::new(
                     types::Union::new(
                         types::Reference::new("boolean", Position::fake()),
@@ -1555,10 +1455,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                type_()
-                    .parse(stream("\\(number)number|none", ""))
-                    .unwrap()
-                    .0,
+                type_(input("\\(number)number|none", "")).unwrap().1,
                 types::Function::new(
                     vec![types::Reference::new("number", Position::fake()).into()],
                     types::Union::new(
@@ -1571,10 +1468,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                type_()
-                    .parse(stream("(\\(number)number)|none", ""))
-                    .unwrap()
-                    .0,
+                type_(input("(\\(number)number)|none", "")).unwrap().1,
                 types::Union::new(
                     types::Function::new(
                         vec![types::Reference::new("number", Position::fake()).into()],
@@ -1590,13 +1484,13 @@ mod tests {
 
         #[test]
         fn parse_reference_type() {
-            assert!(type_().parse(stream("", "")).is_err());
+            assert!(type_(input("", "")).is_err());
             assert_eq!(
-                type_().parse(stream("Foo", "")).unwrap().0,
+                type_(input("Foo", "")).unwrap().1,
                 types::Reference::new("Foo", Position::fake()).into()
             );
             assert_eq!(
-                type_().parse(stream("Foo'Bar", "")).unwrap().0,
+                type_(input("Foo'Bar", "")).unwrap().1,
                 types::Reference::new("Foo'Bar", Position::fake()).into()
             );
         }
@@ -1604,7 +1498,7 @@ mod tests {
         #[test]
         fn parse_list_type() {
             assert_eq!(
-                type_().parse(stream("[number]", "")).unwrap().0,
+                type_(input("[number]", "")).unwrap().1,
                 types::List::new(
                     types::Reference::new("number", Position::fake()),
                     Position::fake()
@@ -1613,7 +1507,7 @@ mod tests {
             );
 
             assert_eq!(
-                type_().parse(stream("[[number]]", "")).unwrap().0,
+                type_(input("[[number]]", "")).unwrap().1,
                 types::List::new(
                     types::List::new(
                         types::Reference::new("number", Position::fake()),
@@ -1625,7 +1519,7 @@ mod tests {
             );
 
             assert_eq!(
-                type_().parse(stream("[number]|[none]", "")).unwrap().0,
+                type_(input("[number]|[none]", "")).unwrap().1,
                 types::Union::new(
                     types::List::new(
                         types::Reference::new("number", Position::fake()),
@@ -1641,7 +1535,7 @@ mod tests {
             );
 
             assert_eq!(
-                type_().parse(stream("\\([number])[none]", "")).unwrap().0,
+                type_(input("\\([number])[none]", "")).unwrap().1,
                 types::Function::new(
                     vec![types::List::new(
                         types::Reference::new("number", Position::fake()),
@@ -1661,7 +1555,7 @@ mod tests {
         #[test]
         fn parse_map_type() {
             assert_eq!(
-                type_().parse(stream("{number:none}", "")).unwrap().0,
+                type_(input("{number:none}", "")).unwrap().1,
                 types::Map::new(
                     types::Reference::new("number", Position::fake()),
                     types::Reference::new("none", Position::fake()),
@@ -1672,15 +1566,15 @@ mod tests {
         }
     }
 
-    mod expressions {
+    mod expression {
         use super::*;
         use pretty_assertions::assert_eq;
 
         #[test]
         fn parse_expression() {
-            assert!(expression().parse(stream("", "")).is_err());
+            assert!(expression(input("", "")).is_err());
             assert_eq!(
-                expression().parse(stream("1", "")).unwrap().0,
+                expression(input("1", "")).unwrap().1,
                 Number::new(
                     NumberRepresentation::FloatingPoint("1".into()),
                     Position::fake()
@@ -1688,11 +1582,11 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                expression().parse(stream("x", "")).unwrap().0,
+                expression(input("x", "")).unwrap().1,
                 Variable::new("x", Position::fake()).into()
             );
             assert_eq!(
-                expression().parse(stream("x + 1", "")).unwrap().0,
+                expression(input("x + 1", "")).unwrap().1,
                 BinaryOperation::new(
                     BinaryOperator::Add,
                     Variable::new("x", Position::fake()),
@@ -1705,7 +1599,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                expression().parse(stream("x + y(z)", "")).unwrap().0,
+                expression(input("x + y(z)", "")).unwrap().1,
                 BinaryOperation::new(
                     BinaryOperator::Add,
                     Variable::new("x", Position::fake()),
@@ -1719,7 +1613,7 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                expression().parse(stream("(x + y)(z)", "")).unwrap().0,
+                expression(input("(x + y)(z)", "")).unwrap().1,
                 Call::new(
                     BinaryOperation::new(
                         BinaryOperator::Add,
@@ -1737,7 +1631,7 @@ mod tests {
         #[test]
         fn parse_deeply_nested_expression() {
             assert_eq!(
-                expression().parse(stream("(((((42)))))", "")).unwrap().0,
+                expression(input("(((((42)))))", "")).unwrap().1,
                 Number::new(
                     NumberRepresentation::FloatingPoint("42".into()),
                     Position::fake()
@@ -1748,9 +1642,9 @@ mod tests {
 
         #[test]
         fn parse_atomic_expression() {
-            assert!(atomic_expression().parse(stream("", "")).is_err());
+            assert!(atomic_expression(input("", "")).is_err());
             assert_eq!(
-                atomic_expression().parse(stream("1", "")).unwrap().0,
+                atomic_expression(input("1", "")).unwrap().1,
                 Number::new(
                     NumberRepresentation::FloatingPoint("1".into()),
                     Position::fake()
@@ -1758,11 +1652,11 @@ mod tests {
                 .into()
             );
             assert_eq!(
-                atomic_expression().parse(stream("x", "")).unwrap().0,
+                atomic_expression(input("x", "")).unwrap().1,
                 Variable::new("x", Position::fake()).into()
             );
             assert_eq!(
-                atomic_expression().parse(stream("(x)", "")).unwrap().0,
+                atomic_expression(input("(x)", "")).unwrap().1,
                 Variable::new("x", Position::fake()).into()
             );
         }
@@ -1770,10 +1664,7 @@ mod tests {
         #[test]
         fn parse_lambda() {
             assert_eq!(
-                lambda()
-                    .parse(stream("\\(x number)number{42}", ""))
-                    .unwrap()
-                    .0,
+                lambda(input("\\(x number)number{42}", "")).unwrap().1,
                 Lambda::new(
                     vec![Argument::new(
                         "x",
@@ -1793,10 +1684,9 @@ mod tests {
             );
 
             assert_eq!(
-                lambda()
-                    .parse(stream("\\(x number,y number)number{42}", ""))
+                lambda(input("\\(x number,y number)number{42}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 Lambda::new(
                     vec![
                         Argument::new("x", types::Reference::new("number", Position::fake())),
@@ -1819,7 +1709,7 @@ mod tests {
         #[test]
         fn parse_lambda_with_reference_type() {
             assert_eq!(
-                lambda().parse(stream("\\() Foo { 42 }", "")).unwrap().0,
+                lambda(input("\\() Foo { 42 }", "")).unwrap().1,
                 Lambda::new(
                     vec![],
                     types::Reference::new("Foo", Position::fake()),
@@ -1839,7 +1729,7 @@ mod tests {
         #[test]
         fn parse_block() {
             assert_eq!(
-                block().parse(stream("{none}", "")).unwrap().0,
+                block(input("{none}", "")).unwrap().1,
                 Block::new(
                     vec![],
                     Variable::new("none", Position::fake()),
@@ -1847,7 +1737,7 @@ mod tests {
                 ),
             );
             assert_eq!(
-                block().parse(stream("{none none}", "")).unwrap().0,
+                block(input("{none none}", "")).unwrap().1,
                 Block::new(
                     vec![Statement::new(
                         None,
@@ -1859,7 +1749,7 @@ mod tests {
                 ),
             );
             assert_eq!(
-                block().parse(stream("{none none none}", "")).unwrap().0,
+                block(input("{none none none}", "")).unwrap().1,
                 Block::new(
                     vec![
                         Statement::new(
@@ -1878,7 +1768,7 @@ mod tests {
                 ),
             );
             assert_eq!(
-                block().parse(stream("{x=none none}", "")).unwrap().0,
+                block(input("{x=none none}", "")).unwrap().1,
                 Block::new(
                     vec![Statement::new(
                         Some("x".into()),
@@ -1890,7 +1780,7 @@ mod tests {
                 ),
             );
             assert_eq!(
-                block().parse(stream("{x==x}", "")).unwrap().0,
+                block(input("{x==x}", "")).unwrap().1,
                 Block::new(
                     vec![],
                     BinaryOperation::new(
@@ -1907,7 +1797,7 @@ mod tests {
         #[test]
         fn parse_statement() {
             assert_eq!(
-                statement().parse(stream("x==x", "")).unwrap().0,
+                statement(input("x==x", "")).unwrap().1,
                 Statement::new(
                     None,
                     BinaryOperation::new(
@@ -1924,10 +1814,7 @@ mod tests {
         #[test]
         fn parse_if() {
             assert_eq!(
-                if_()
-                    .parse(stream("if true { 42 } else { 13 }", ""))
-                    .unwrap()
-                    .0,
+                if_(input("if true { 42 } else { 13 }", "")).unwrap().1,
                 If::new(
                     vec![IfBranch::new(
                         Variable::new("true", Position::fake()),
@@ -1952,10 +1839,9 @@ mod tests {
                 )
             );
             assert_eq!(
-                if_()
-                    .parse(stream("if if true {true}else{true}{42}else{13}", ""))
+                if_(input("if if true {true}else{true}{42}else{13}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 If::new(
                     vec![IfBranch::new(
                         If::new(
@@ -1995,10 +1881,9 @@ mod tests {
                 )
             );
             assert_eq!(
-                if_()
-                    .parse(stream("if true {1}else if true {2}else{3}", ""))
+                if_(input("if true {1}else if true {2}else{3}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 If::new(
                     vec![
                         IfBranch::new(
@@ -2040,10 +1925,7 @@ mod tests {
         #[test]
         fn parse_if_with_equal_operator() {
             assert_eq!(
-                expression()
-                    .parse(stream("if x==y {none}else{none}", ""))
-                    .unwrap()
-                    .0,
+                expression(input("if x==y {none}else{none}", "")).unwrap().1,
                 If::new(
                     vec![IfBranch::new(
                         BinaryOperation::new(
@@ -2072,10 +1954,9 @@ mod tests {
         #[test]
         fn parse_if_type() {
             assert_eq!(
-                if_type()
-                    .parse(stream("if x=y as boolean {none}else{none}", ""))
+                if_type(input("if x=y as boolean {none}else{none}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 IfType::new(
                     "x",
                     Variable::new("y", Position::fake()),
@@ -2097,13 +1978,12 @@ mod tests {
             );
 
             assert_eq!(
-                if_type()
-                    .parse(stream(
-                        "if x=y as boolean{none}else if none{none}else{none}",
-                        ""
-                    ))
-                    .unwrap()
-                    .0,
+                if_type(input(
+                    "if x=y as boolean{none}else if none{none}else{none}",
+                    ""
+                ))
+                .unwrap()
+                .1,
                 IfType::new(
                     "x",
                     Variable::new("y", Position::fake()),
@@ -2135,10 +2015,9 @@ mod tests {
             );
 
             assert_eq!(
-                if_type()
-                    .parse(stream("if x=y as boolean{none}else if none{none}", ""))
+                if_type(input("if x=y as boolean{none}else if none{none}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 IfType::new(
                     "x",
                     Variable::new("y", Position::fake()),
@@ -2169,10 +2048,9 @@ mod tests {
         #[test]
         fn parse_if_list() {
             assert_eq!(
-                if_list()
-                    .parse(stream("if[x,...xs]=xs {none}else{none}", ""))
+                if_list(input("if[x,...xs]=xs {none}else{none}", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 IfList::new(
                     Variable::new("xs", Position::fake()),
                     "x",
@@ -2195,10 +2073,7 @@ mod tests {
         #[test]
         fn parse_if_map() {
             assert_eq!(
-                if_map()
-                    .parse(stream("if x=xs[42]{none}else{none}", ""))
-                    .unwrap()
-                    .0,
+                if_map(input("if x=xs[42]{none}else{none}", "")).unwrap().1,
                 IfMap::new(
                     "x",
                     Variable::new("xs", Position::fake()),
@@ -2228,7 +2103,12 @@ mod tests {
             #[test]
             fn parse_call() {
                 assert_eq!(
-                    expression().parse(stream("f()", "")).unwrap().0,
+                    expression(input("f ()", "")).unwrap().1,
+                    Variable::new("f", Position::fake()).into()
+                );
+
+                assert_eq!(
+                    expression(input("f()", "")).unwrap().1,
                     Call::new(
                         Variable::new("f", Position::fake()),
                         vec![],
@@ -2238,7 +2118,7 @@ mod tests {
                 );
 
                 assert_eq!(
-                    expression().parse(stream("f()()", "")).unwrap().0,
+                    expression(input("f()()", "")).unwrap().1,
                     Call::new(
                         Call::new(
                             Variable::new("f", Position::fake()),
@@ -2252,7 +2132,7 @@ mod tests {
                 );
 
                 assert_eq!(
-                    expression().parse(stream("f(1)", "")).unwrap().0,
+                    expression(input("f(1)", "")).unwrap().1,
                     Call::new(
                         Variable::new("f", Position::fake()),
                         vec![Number::new(
@@ -2266,7 +2146,7 @@ mod tests {
                 );
 
                 assert_eq!(
-                    expression().parse(stream("f(1,)", "")).unwrap().0,
+                    expression(input("f(1,)", "")).unwrap().1,
                     Call::new(
                         Variable::new("f", Position::fake()),
                         vec![Number::new(
@@ -2280,7 +2160,7 @@ mod tests {
                 );
 
                 assert_eq!(
-                    expression().parse(stream("f(1, 2)", "")).unwrap().0,
+                    expression(input("f(1, 2)", "")).unwrap().1,
                     Call::new(
                         Variable::new("f", Position::fake()),
                         vec![
@@ -2301,7 +2181,7 @@ mod tests {
                 );
 
                 assert_eq!(
-                    expression().parse(stream("f(1, 2,)", "")).unwrap().0,
+                    expression(input("f(1, 2,)", "")).unwrap().1,
                     Call::new(
                         Variable::new("f", Position::fake()),
                         vec![
@@ -2324,19 +2204,18 @@ mod tests {
 
             #[test]
             fn fail_to_parse_call() {
-                let source = "f(1+)";
-
-                insta::assert_debug_snapshot!(expression()
-                    .parse(stream(source, ""))
-                    .map_err(|error| ParseError::new(source, "", error))
-                    .err());
+                // TODO Should we rather commit to parse calls?
+                assert_eq!(
+                    expression(input("f(1+)", "")).unwrap().1,
+                    Variable::new("f", Position::fake()).into()
+                );
             }
         }
 
         #[test]
         fn parse_try_operation() {
             assert_eq!(
-                expression().parse(stream("x?", "")).unwrap().0,
+                expression(input("x?", "")).unwrap().1,
                 UnaryOperation::new(
                     UnaryOperator::Try,
                     Variable::new("x", Position::fake()),
@@ -2347,8 +2226,8 @@ mod tests {
         }
 
         #[test]
-        fn parse_unary_operation() {
-            assert!(prefix_operation().parse(stream("", "")).is_err());
+        fn parse_prefix_operation() {
+            assert!(prefix_operation(input("", "")).is_err());
 
             for (source, expected) in &[
                 (
@@ -2417,19 +2296,16 @@ mod tests {
                     ),
                 ),
             ] {
-                assert_eq!(
-                    prefix_operation().parse(stream(source, "")).unwrap().0,
-                    *expected
-                );
+                assert_eq!(prefix_operation(input(source, "")).unwrap().1, *expected);
             }
         }
 
         #[test]
         fn parse_prefix_operator() {
-            assert!(prefix_operator().parse(stream("", "")).is_err());
+            assert!(prefix_operator(input("", "")).is_err());
 
             assert_eq!(
-                prefix_operator().parse(stream("!", "")).unwrap().0,
+                prefix_operator(input("!", "")).unwrap().1,
                 UnaryOperator::Not
             );
         }
@@ -2684,14 +2560,14 @@ mod tests {
                     .into(),
                 ),
             ] {
-                assert_eq!(expression().parse(stream(source, "")).unwrap().0, target);
+                assert_eq!(expression(input(source, "")).unwrap().1, target);
             }
         }
 
         #[test]
         fn parse_binary_operator() {
-            assert!(binary_operator().parse(stream("", "")).is_err());
-            assert!(binary_operator().parse(stream("++", "")).is_err());
+            assert!(binary_operator(input("", "")).is_err());
+            assert!(binary_operator(input("+=", "")).is_err());
 
             for (source, expected) in &[
                 ("+", BinaryOperator::Add),
@@ -2707,24 +2583,21 @@ mod tests {
                 ("&", BinaryOperator::And),
                 ("|", BinaryOperator::Or),
             ] {
-                assert_eq!(
-                    binary_operator().parse(stream(source, "")).unwrap().0,
-                    *expected
-                );
+                assert_eq!(binary_operator(input(source, "")).unwrap().1, *expected);
             }
         }
 
         #[test]
         fn parse_record() {
-            assert!(record().parse(stream("Foo", "")).is_err());
+            assert!(record(input("Foo", "")).is_err());
 
             assert_eq!(
-                record().parse(stream("Foo{}", "")).unwrap().0,
+                record(input("Foo{}", "")).unwrap().1,
                 Record::new("Foo", None, vec![], Position::fake())
             );
 
             assert_eq!(
-                expression().parse(stream("Foo{foo:42}", "")).unwrap().0,
+                expression(input("Foo{foo:42}", "")).unwrap().1,
                 Record::new(
                     "Foo",
                     None,
@@ -2742,7 +2615,7 @@ mod tests {
             );
 
             assert_eq!(
-                record().parse(stream("Foo{foo:42}", "")).unwrap().0,
+                record(input("Foo{foo:42}", "")).unwrap().1,
                 Record::new(
                     "Foo",
                     None,
@@ -2759,7 +2632,7 @@ mod tests {
             );
 
             assert_eq!(
-                record().parse(stream("Foo{foo:42,bar:42}", "")).unwrap().0,
+                record(input("Foo{foo:42,bar:42}", "")).unwrap().1,
                 Record::new(
                     "Foo",
                     None,
@@ -2785,13 +2658,8 @@ mod tests {
                 )
             );
 
-            assert!(record().parse(stream("Foo{foo:42,foo:42}", "")).is_err());
-
             assert_eq!(
-                expression()
-                    .parse(stream("foo(Foo{foo:42})", ""))
-                    .unwrap()
-                    .0,
+                expression(input("foo(Foo{foo:42})", "")).unwrap().1,
                 Call::new(
                     Variable::new("foo", Position::fake()),
                     vec![Record::new(
@@ -2814,7 +2682,7 @@ mod tests {
             );
 
             assert_eq!(
-                record().parse(stream("Foo{foo:bar(42)}", "")).unwrap().0,
+                record(input("Foo{foo:bar(42)}", "")).unwrap().1,
                 Record::new(
                     "Foo",
                     None,
@@ -2835,10 +2703,10 @@ mod tests {
                 )
             );
 
-            assert!(record().parse(stream("Foo{...foo,}", "")).is_err());
+            assert!(record(input("Foo{...foo,}", "")).is_err());
 
             assert_eq!(
-                record().parse(stream("Foo{...foo,bar:42}", "")).unwrap().0,
+                record(input("Foo{...foo,bar:42}", "")).unwrap().1,
                 Record::new(
                     "Foo",
                     Some(Variable::new("foo", Position::fake()).into()),
@@ -2855,7 +2723,7 @@ mod tests {
             );
 
             assert_eq!(
-                record().parse(stream("Foo{...foo,bar:42,}", "")).unwrap().0,
+                record(input("Foo{...foo,bar:42,}", "")).unwrap().1,
                 Record::new(
                     "Foo",
                     Some(Variable::new("foo", Position::fake()).into()),
@@ -2872,10 +2740,7 @@ mod tests {
             );
 
             assert_eq!(
-                expression()
-                    .parse(stream("Foo{...foo,bar:42}", ""))
-                    .unwrap()
-                    .0,
+                expression(input("Foo{...foo,bar:42}", "")).unwrap().1,
                 Record::new(
                     "Foo",
                     Some(Variable::new("foo", Position::fake()).into()),
@@ -2892,39 +2757,31 @@ mod tests {
                 .into(),
             );
 
-            assert!(record().parse(stream("Foo{...foo}", "")).is_err());
-            assert!(record()
-                .parse(stream("Foo{...foo,bar:42,bar:42}", ""))
-                .is_err());
-            assert!(record().parse(stream("Foo{...(foo),bar:42}", "")).is_ok());
-            assert!(record()
-                .parse(stream("Foo{...foo(bar),bar:42}", ""))
-                .is_ok());
-            assert!(record()
-                .parse(stream("Foo{...if true { none } else { none },bar:42}", ""))
-                .is_ok());
+            assert!(record(input("Foo{...foo}", "")).is_err());
+            assert!(record(input("Foo{...(foo),bar:42}", "")).is_ok());
+            assert!(record(input("Foo{...foo(bar),bar:42}", "")).is_ok());
+            assert!(record(input("Foo{...if true { none } else { none },bar:42}", "")).is_ok());
         }
 
         #[test]
         fn parse_variable() {
-            assert!(variable().parse(stream("", "")).is_err());
+            assert!(variable(input("", "")).is_err());
 
             assert_eq!(
-                variable().parse(stream("x", "")).unwrap().0,
+                variable(input("x", "")).unwrap().1,
                 Variable::new("x", Position::fake()),
             );
 
             assert_eq!(
-                variable().parse(stream("Foo.x", "")).unwrap().0,
+                variable(input("Foo.x", "")).unwrap().1,
                 Variable::new("Foo", Position::fake()),
             );
         }
 
         #[test]
         fn parse_number_literal() {
-            assert!(number_literal().parse(stream("", "")).is_err());
-            assert!(number_literal().parse(stream("foo", "")).is_err());
-            assert!(number_literal().parse(stream("01", "")).is_err());
+            assert!(number_literal(input("", "")).is_err());
+            assert!(number_literal(input("foo", "")).is_err());
 
             for (source, value) in [
                 ("0", NumberRepresentation::FloatingPoint("0".into())),
@@ -2943,7 +2800,7 @@ mod tests {
                 ("0xfa", NumberRepresentation::Hexadecimal("fa".into())),
             ] {
                 assert_eq!(
-                    number_literal().parse(stream(source, "")).unwrap().0,
+                    number_literal(input(source, "")).unwrap().1,
                     Number::new(value, Position::fake())
                 );
             }
@@ -2951,12 +2808,14 @@ mod tests {
 
         #[test]
         fn parse_string_literal() {
-            assert!(string_literal().parse(stream("", "")).is_err());
-            assert!(string_literal().parse(stream("foo", "")).is_err());
+            assert!(string_literal(input("", "")).is_err());
+            assert!(string_literal(input("foo", "")).is_err());
+            assert!(string_literal(input("\\a", "")).is_err());
 
             for (source, value) in &[
                 (r#""""#, ""),
                 (r#""foo""#, "foo"),
+                (r#" "foo""#, "foo"),
                 (r#""foo bar""#, "foo bar"),
                 (r#""\"""#, "\\\""),
                 (r#""\n""#, "\\n"),
@@ -2967,7 +2826,7 @@ mod tests {
                 (r#""\n\n""#, "\\n\\n"),
             ] {
                 assert_eq!(
-                    string_literal().parse(stream(source, "")).unwrap().0,
+                    string_literal(input(source, "")).unwrap().1,
                     ByteString::new(*value, Position::fake())
                 );
             }
@@ -3091,10 +2950,7 @@ mod tests {
                     ),
                 ),
             ] {
-                assert_eq!(
-                    expression().parse(stream(source, "")).unwrap().0,
-                    target.into()
-                );
+                assert_eq!(expression(input(source, "")).unwrap().1, target.into());
             }
         }
 
@@ -3131,7 +2987,7 @@ mod tests {
                 ),
             ] {
                 assert_eq!(
-                    list_comprehension().parse(stream(source, "")).unwrap().0,
+                    list_comprehension(input(source, "")).unwrap().1,
                     target.into()
                 );
             }
@@ -3219,17 +3075,16 @@ mod tests {
                     .into(),
                 ),
             ] {
-                assert_eq!(expression().parse(stream(source, "")).unwrap().0, target);
+                assert_eq!(expression(input(source, "")).unwrap().1, target);
             }
         }
 
         #[test]
         fn parse_map_iteration_comprehension() {
             assert_eq!(
-                list_comprehension()
-                    .parse(stream("[none v for k, v in xs]", ""))
+                list_comprehension(input("[none v for k, v in xs]", ""))
                     .unwrap()
-                    .0,
+                    .1,
                 MapIterationComprehension::new(
                     types::Reference::new("none", Position::fake()),
                     Variable::new("v", Position::fake()),
@@ -3245,55 +3100,49 @@ mod tests {
 
     #[test]
     fn parse_identifier() {
-        assert!(identifier().parse(stream("if", "")).is_err());
-        assert!(identifier().parse(stream("1foo", "")).is_err());
+        assert!(identifier(input("if", "")).is_err());
+        assert!(identifier(input("1foo", "")).is_err());
+        assert_eq!(identifier(input("foo", "")).unwrap().1, "foo".to_string());
         assert_eq!(
-            identifier().parse(stream("foo", "")).unwrap().0,
-            "foo".to_string()
-        );
-        assert_eq!(
-            identifier().parse(stream("foo42", "")).unwrap().0,
+            identifier(input("foo42", "")).unwrap().1,
             "foo42".to_string()
         );
     }
 
     #[test]
     fn parse_keyword() {
-        assert!(keyword("type").parse(stream("bar", "")).is_err());
+        assert!(keyword("type").parse(input("bar", "")).is_err());
         // spell-checker: disable-next-line
-        assert!(keyword("type").parse(stream("typer", "")).is_err());
-        assert!(keyword("type").parse(stream("type_", "")).is_err());
-        assert!(keyword("type").parse(stream("type", "")).is_ok());
+        assert!(keyword("type").parse(input("typer", "")).is_err());
+        assert!(keyword("type").parse(input("type_", "")).is_err());
+        assert!(keyword("type").parse(input("type", "")).is_ok());
     }
 
     #[test]
     fn parse_sign() {
-        assert!(sign("+").parse(stream("", "")).is_err());
-        assert!(sign("+").parse(stream("-", "")).is_err());
-        assert!(sign("+").parse(stream("+", "")).is_ok());
-        assert!(sign("++").parse(stream("++", "")).is_ok());
-        assert!(sign("+").parse(stream("++", "")).is_err());
-    }
-
-    #[test]
-    fn parse_position() {
-        assert!(position().parse(stream("", "")).is_ok());
+        assert!(sign("+")(input("", "")).is_err());
+        assert!(sign("+")(input("-", "")).is_err());
+        assert!(sign("+")(input("+", "")).is_ok());
+        assert!(sign("++")(input("++", "")).is_ok());
+        assert!(sign("+")(input("++", "")).is_ok());
+        assert!(sign("+")(input("+=", "")).is_err());
+        assert!(sign("\\")(input("\\", "")).is_ok());
     }
 
     #[test]
     fn parse_blank() {
-        assert!(blank().with(eof()).parse(stream(" ", "")).is_ok());
-        assert!(blank().with(eof()).parse(stream("\n", "")).is_ok());
-        assert!(blank().with(eof()).parse(stream(" \n", "")).is_ok());
-        assert!(blank().with(eof()).parse(stream("\t", "")).is_ok());
-        assert!(blank().with(eof()).parse(stream("# foo", "")).is_ok());
+        assert!(all_consuming(blank)(input(" ", "")).is_ok());
+        assert!(all_consuming(blank)(input("\n", "")).is_ok());
+        assert!(all_consuming(blank)(input(" \n", "")).is_ok());
+        assert!(all_consuming(blank)(input("\t", "")).is_ok());
+        assert!(all_consuming(blank)(input("# foo", "")).is_ok());
     }
 
     #[test]
     fn parse_comment() {
-        assert!(comment().parse(stream("#", "")).is_ok());
-        assert!(comment().parse(stream("#\n", "")).is_ok());
-        assert!(comment().parse(stream("#x\n", "")).is_ok());
+        assert!(comment(input("#", "")).is_ok());
+        assert!(comment(input("#\n", "")).is_ok());
+        assert!(comment(input("#x\n", "")).is_ok());
     }
 
     mod comments {
@@ -3303,7 +3152,7 @@ mod tests {
         #[test]
         fn parse_comment() {
             assert_eq!(
-                comments().parse(stream("#foo", "")).unwrap().0,
+                comments(input("#foo", "")).unwrap().1,
                 vec![Comment::new("foo", Position::fake())]
             );
         }
@@ -3311,7 +3160,7 @@ mod tests {
         #[test]
         fn parse_comment_after_space() {
             assert_eq!(
-                comments().parse(stream(" #foo", "")).unwrap().0,
+                comments(input(" #foo", "")).unwrap().1,
                 vec![Comment::new("foo", Position::fake())]
             );
         }
@@ -3319,7 +3168,7 @@ mod tests {
         #[test]
         fn parse_comment_before_space() {
             assert_eq!(
-                comments().parse(stream("#foo\n #bar", "")).unwrap().0,
+                comments(input("#foo\n #bar", "")).unwrap().1,
                 vec![
                     Comment::new("foo", Position::fake()),
                     Comment::new("bar", Position::fake())
@@ -3330,7 +3179,7 @@ mod tests {
         #[test]
         fn parse_comment_before_newlines() {
             assert_eq!(
-                comments().parse(stream("#foo\n\n", "")).unwrap().0,
+                comments(input("#foo\n\n", "")).unwrap().1,
                 vec![Comment::new("foo", Position::fake())]
             );
         }
@@ -3338,18 +3187,17 @@ mod tests {
         #[test]
         fn parse_two_line_comments() {
             assert_eq!(
-                comments()
-                    .parse(stream(
-                        indoc!(
-                            "
+                comments(input(
+                    indoc!(
+                        "
                             #foo
                             #bar
                             "
-                        ),
-                        ""
-                    ))
-                    .unwrap()
-                    .0,
+                    ),
+                    ""
+                ))
+                .unwrap()
+                .1,
                 vec![
                     Comment::new("foo", Position::fake()),
                     Comment::new("bar", Position::fake())
@@ -3360,7 +3208,7 @@ mod tests {
         #[test]
         fn parse_comment_after_identifier() {
             assert_eq!(
-                comments().parse(stream("foo#foo", "")).unwrap().0,
+                comments(input("foo#foo", "")).unwrap().1,
                 vec![Comment::new("foo", Position::fake())]
             );
         }
@@ -3368,7 +3216,7 @@ mod tests {
         #[test]
         fn parse_comment_before_identifier() {
             assert_eq!(
-                comments().parse(stream("#foo\nfoo#bar", "")).unwrap().0,
+                comments(input("#foo\nfoo#bar", "")).unwrap().1,
                 vec![
                     Comment::new("foo", Position::fake()),
                     Comment::new("bar", Position::fake())
@@ -3379,7 +3227,7 @@ mod tests {
         #[test]
         fn parse_comment_after_keyword() {
             assert_eq!(
-                comments().parse(stream("if#foo", "")).unwrap().0,
+                comments(input("if#foo", "")).unwrap().1,
                 vec![Comment::new("foo", Position::fake())]
             );
         }
@@ -3387,7 +3235,7 @@ mod tests {
         #[test]
         fn parse_comment_before_keyword() {
             assert_eq!(
-                comments().parse(stream("#foo\nif#bar", "")).unwrap().0,
+                comments(input("#foo\nif#bar", "")).unwrap().1,
                 vec![
                     Comment::new("foo", Position::fake()),
                     Comment::new("bar", Position::fake())
@@ -3398,7 +3246,7 @@ mod tests {
         #[test]
         fn parse_comment_after_sign() {
             assert_eq!(
-                comments().parse(stream("+#foo", "")).unwrap().0,
+                comments(input("+#foo", "")).unwrap().1,
                 vec![Comment::new("foo", Position::fake())]
             );
         }
@@ -3406,7 +3254,7 @@ mod tests {
         #[test]
         fn parse_comment_before_sign() {
             assert_eq!(
-                comments().parse(stream("#foo\n+#bar", "")).unwrap().0,
+                comments(input("#foo\n+#bar", "")).unwrap().1,
                 vec![
                     Comment::new("foo", Position::fake()),
                     Comment::new("bar", Position::fake())
@@ -3417,7 +3265,7 @@ mod tests {
         #[test]
         fn parse_comment_after_string() {
             assert_eq!(
-                comments().parse(stream("\"string\"#foo", "")).unwrap().0,
+                comments(input("\"string\"#foo", "")).unwrap().1,
                 vec![Comment::new("foo", Position::fake())]
             );
         }
@@ -3425,10 +3273,7 @@ mod tests {
         #[test]
         fn parse_comment_before_string() {
             assert_eq!(
-                comments()
-                    .parse(stream("#foo\n\"string\"#bar", ""))
-                    .unwrap()
-                    .0,
+                comments(input("#foo\n\"string\"#bar", "")).unwrap().1,
                 vec![
                     Comment::new("foo", Position::fake()),
                     Comment::new("bar", Position::fake())
